feat(auth): implementar permisos en users, permisos y roles

+++ auth-service/src/graphql/resolvers/permission.resolver.js
@@ -1,11 +1,12 @@
 const Permission = require("../../models/permission.model");
-
+const { requirePermission } = require("../../helper");
 
 
 module.exports = {
 
-    getPermissions: async () => {
+    getPermissions: async (args, context) => {
         try {
+            requirePermission(context.user, "getPermissions");
             const permissions = await Permission.find();
             if (!permissions) {
                 return [];
@@ -15,8 +16,9 @@ module.exports = {
             throw new Error("Error fetching permissions: " + error.message);
         }
     },
-    getPermission: async ({id}) =>{
+    getPermission: async ({id}, context) =>{
         try {
+            requirePermission(context.user, "getPermission");
             const permission = await Permission.findById(id);
             if(!permission){
                 
@@ -27,9 +29,10 @@ module.exports = {
             throw new Error("Error al obtener el permiso: "+ error.message);
         }
     },
-
-    createPermission: async (args)=>{
+    
+    createPermission: async (args, context)=>{
         try {
+            requirePermission(context.user, "createPermission");
             const permission = await Permission.create(args);
             if (!permission) {
                 throw new Error("Error creating permission");
@@ -39,9 +42,10 @@ module.exports = {
             throw new Error("Error creating permission: " + error.message);
         }
     },
-
-    updatePermission: async (args) => {
+    
+    updatePermission: async (args,context) => {
         try {
+            requirePermission(context.user, "updatePermission");
             const permission = await Permission.findByIdAndUpdate(
                 args.id,
                 { name: args.name, description: args.description, route: args.route },
@@ -56,8 +60,9 @@ module.exports = {
         }
     },
 
-    deletePermission: async (args) =>{
+    deletePermission: async (args,context) =>{
         try {
+            requirePermission(context.user, "deletePermission");
             const permission = await Permission.findByIdAndDelete(args.id);
             if (!permission) {
                 throw new Error("Permission not found");
